-Turkish-People-s-Attitudes-Towards-E-Commerce-Survey-Analysis
In this project, the aim is to draw a conclusion about the attitudes of Turkish people towards e-commerce. That's why, a survey releated to topic is constructed and applied with participants by snowball method. After collecting data, by applying some appropriate statistical tests and visualizations, the researchers reached a conclusion. Codes are at the end